# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Yaaran E Ilm API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: tutors
    description: Tutor operations
  - name: videos
    description: Video operations
  - name: classes
    description: Class/session operations
  - name: stats
    description: Platform statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tutors:
    get:
      operationId: listTutors
      tags: [tutors]
      summary: List all tutors
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tutors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tutor"

  /tutors/{id}:
    get:
      operationId: getTutor
      tags: [tutors]
      summary: Get tutor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tutor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tutor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTutor
      tags: [tutors]
      summary: Update tutor profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutorUpdate"
      responses:
        "200":
          description: Updated tutor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tutor"

  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List all educational videos
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
        - name: uploaderId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
    post:
      operationId: createVideo
      tags: [videos]
      summary: Upload/create a video entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoInput"
      responses:
        "201":
          description: Video created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"

  /videos/featured:
    get:
      operationId: getFeaturedVideos
      tags: [videos]
      summary: Get featured/popular videos
      responses:
        "200":
          description: Featured videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"

  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get video by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteVideo
      tags: [videos]
      summary: Delete a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List available classes
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
        - name: tutorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a class (tutors only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"

  /classes/{id}:
    get:
      operationId: getClass
      tags: [classes]
      summary: Get class by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /classes/{id}/enroll:
    post:
      operationId: enrollInClass
      tags: [classes]
      summary: Enroll in a class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enrolled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /my/classes:
    get:
      operationId: getMyClasses
      tags: [classes]
      summary: Get classes for current user (enrolled or created)
      responses:
        "200":
          description: My classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"

  /my/videos:
    get:
      operationId: getMyVideos
      tags: [videos]
      summary: Get videos uploaded by current user
      responses:
        "200":
          description: My videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"

  /stats:
    get:
      operationId: getPlatformStats
      tags: [stats]
      summary: Get platform-wide stats
      responses:
        "200":
          description: Platform statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, tutor]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt

    RegisterInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [student, tutor]
        subject:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        level:
          type: ["string", "null"]
      required:
        - name
        - email
        - password
        - role

    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    Tutor:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        level:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        totalStudents:
          type: integer
        totalVideos:
          type: integer
        totalClasses:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - name
        - email
        - totalStudents
        - totalVideos
        - totalClasses
        - createdAt

    TutorUpdate:
      type: object
      properties:
        bio:
          type: string
        subject:
          type: string
        level:
          type: string
        avatarUrl:
          type: string

    Video:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        videoUrl:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        level:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        views:
          type: integer
        uploaderId:
          type: integer
        uploaderName:
          type: string
        uploaderRole:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - videoUrl
        - views
        - uploaderId
        - uploaderName
        - uploaderRole
        - createdAt

    VideoInput:
      type: object
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
        videoUrl:
          type: string
        thumbnailUrl:
          type: string
        subject:
          type: string
        level:
          type: string
        duration:
          type: string
      required:
        - title
        - videoUrl

    Class:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        level:
          type: ["string", "null"]
        tutorId:
          type: integer
        tutorName:
          type: string
        tutorAvatarUrl:
          type: ["string", "null"]
        schedule:
          type: ["string", "null"]
        maxStudents:
          type: integer
        enrolledCount:
          type: integer
        price:
          type: ["number", "null"]
        isFree:
          type: boolean
        status:
          type: string
          enum: [open, full, closed]
        createdAt:
          type: string
      required:
        - id
        - title
        - tutorId
        - tutorName
        - maxStudents
        - enrolledCount
        - isFree
        - status
        - createdAt

    ClassInput:
      type: object
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
        subject:
          type: string
        level:
          type: string
        schedule:
          type: string
        maxStudents:
          type: integer
        price:
          type: number
        isFree:
          type: boolean
      required:
        - title
        - maxStudents
        - isFree

    PlatformStats:
      type: object
      properties:
        totalTutors:
          type: integer
        totalStudents:
          type: integer
        totalVideos:
          type: integer
        totalClasses:
          type: integer
        subjectBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/SubjectCount"
        levelBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/LevelCount"
      required:
        - totalTutors
        - totalStudents
        - totalVideos
        - totalClasses
        - subjectBreakdown
        - levelBreakdown

    SubjectCount:
      type: object
      properties:
        subject:
          type: string
        count:
          type: integer
      required:
        - subject
        - count

    LevelCount:
      type: object
      properties:
        level:
          type: string
        count:
          type: integer
      required:
        - level
        - count