# lib/api-zod/src/generated/types/classInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yaaran E Ilm API
 * OpenAPI spec version: 0.1.0
 */

export interface ClassInput {
  /** @minLength 3 */
  title: string;
  description?: string;
  subject?: string;
  level?: string;
  schedule?: string;
  maxStudents: number;
  price?: number;
  isFree: boolean;
}