# lib/api-zod/src/generated/types/tutor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yaaran E Ilm API
 * OpenAPI spec version: 0.1.0
 */

export interface Tutor {
  id: number;
  userId: number;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  level?: string | null;
  /** @nullable */
  rating?: number | null;
  totalStudents: number;
  totalVideos: number;
  totalClasses: number;
  createdAt: string;
}